Create User Interface design
This is the code to create the user interface for the user to interact with the application.It displays the most famous candidates Hillary Clinton and Donald Trump.The user can then click on the button with each candidate's name and then interact. The user can also view the fiery debate that happened between the two candidates.